airflow-greenplum/airflow/requirements.txt
confluent-kafka==2.3.0
psycopg2-binary==2.9.9
pandas==2.2.2
